Update API To For Additional Device Operations
Devices will be retrieved by name. Device ID will not be stored by client, but will be return via API call using the Device Name.

'Device List' table was updated to add a field for "Location" and an API call was added to allow this field to be update.

A table with connection information for the Mosquitto broker was created. The broker's ip can be retrieved via API call.